chore: register service worker outside of main bundle
