+++ packages/board/public/index.html
@@ -25,5 +25,6 @@
 <body>
   <script type="application/json+config"></script>
   <script async src="bundle.js"></script>
+  <script async src="register-service-worker.js"></script>
 </body>
 </html>
