+++ packages/board/src/main.js
@@ -1,7 +1,3 @@
-import {
-  debounce
-} from 'min-dash';
-
 import { mount } from 'svelte';
 
 import Taskboard from './Taskboard.svelte';
@@ -15,26 +11,5 @@ const taskboard = mount(Taskboard, {
 });
 
 console.timeEnd('Taskboard#create');
-
-if ('serviceWorker' in navigator) {
-
-  const triggerReload = debounce(() => {
-    window.location.reload();
-  }, 1500);
-
-  window.addEventListener('load', () => {
-    navigator.serviceWorker.register('/service-worker.js').catch((err) => {
-      console.warn('Failed to register service worker', err);
-    });
-
-    navigator.serviceWorker.addEventListener('message', event => {
-      if (event.data.message === 'resource.changed') {
-        console.log('Resource changed', event.data.url);
-
-        triggerReload();
-      }
-    });
-  });
-}
 
 export default taskboard;
